+++ src/app/app.module.ts
@@ -4,9 +4,6 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms'
 import { BrowserModule } from '@angular/platform-browser'
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations'
 
-import { MatDialogModule } from '@angular/material/dialog'
-import { MatSnackBarModule } from '@angular/material/snack-bar'
-
 import { AngularFireModule, FirebaseOptionsToken } from '@angular/fire'
 import { AngularFireAuthModule } from '@angular/fire/auth'
 import { AngularFirestoreModule } from '@angular/fire/firestore'
@@ -20,7 +17,6 @@ import { CoreModule } from '@core/core.module'
 import { FirestoreService } from '@core/firestore.service'
 import { FooterComponent } from '@core/footer/footer.component'
 
-import { AdminComponent } from '@modules/admin/admin.component'
 import { LeaderboardComponent } from '@modules/leaderboard/leaderboard.component'
 import { LoginComponent } from '@modules/login/login.component'
 import { UserProfileComponent } from '@modules/user-profile/user-profile.component'
@@ -28,13 +24,13 @@ import { UserProfileComponent } from '@modules/user-profile/user-profile.compone
 import { Environment } from '@environments/environment'
 import { Ng2ImgToolsModule } from 'ng2-img-tools'
 
+import { MatDialogModule } from '@angular/material/dialog'
+import { MatSnackBarModule } from '@angular/material/snack-bar'
 import { ScrollingModule } from '@angular/cdk/scrolling'
 import { MatButtonModule } from '@angular/material/button'
-import { MatTabsModule } from '@angular/material/tabs'
 
 @NgModule({
   declarations: [
-    AdminComponent,
     AppComponent,
     FooterComponent,
     LeaderboardComponent,
@@ -57,7 +53,6 @@ import { MatTabsModule } from '@angular/material/tabs'
     Ng2ImgToolsModule,
     ReactiveFormsModule,
     MatButtonModule,
-    MatTabsModule,
     ScrollingModule
   ],
   providers: [