+++ src/app/modules/dashboard/dashboard.module.ts
@@ -1,0 +1,27 @@
+import { NgModule } from '@angular/core'
+import { CommonModule } from '@angular/common'
+import { DashboardComponent } from './dashboard.component'
+
+import { FormsModule, ReactiveFormsModule } from '@angular/forms'
+
+import { ScrollingModule } from '@angular/cdk/scrolling'
+import { MatTabsModule } from '@angular/material/tabs'
+import { MatDialogModule } from '@angular/material/dialog'
+
+import { Routes, RouterModule } from '@angular/router'
+
+const routes: Routes = [{ path: '', component: DashboardComponent }]
+
+@NgModule({
+  declarations: [DashboardComponent],
+  imports: [
+    CommonModule,
+    RouterModule.forChild(routes),
+    ScrollingModule,
+    MatTabsModule,
+    MatDialogModule,
+    FormsModule,
+    ReactiveFormsModule
+  ]
+})
+export class DashboardModule {}